split up space components into own components
