+++ src/pageContent/explore/Asteroids.module.css
@@ -1,0 +1,90 @@
+.asteroids {
+  flex-basis: 10%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 0 0.5rem;
+  box-sizing: border-box;
+  border: 2px solid transparent;
+  transition: 0.5s ease-in-out;
+
+  &:hover {
+    border: 2px solid #ffffff;
+    border-top: 2px solid transparent;
+  }
+
+  img {
+    filter: sepia(1) hue-rotate(5deg);
+  }
+
+  img:nth-child(1) {
+    width: 55px;
+    transform: translateX(-20px) translateY(21px);
+  }
+  img:nth-child(2) {
+    width: 65px;
+    transform: translateY(14px) rotate(220deg);
+  }
+  img:nth-child(3) {
+    width: 80px;
+    transform: translateX(10px) translateY(7px) rotate(360deg);
+  }
+  img:nth-child(4) {
+    width: 70px;
+    transform: translateX(5px) rotate(230deg);
+  }
+  img:nth-child(5) {
+    width: 60px;
+    transform: translateY(-7px) rotate(330deg);
+  }
+  img:nth-child(6) {
+    width: 50px;
+    transform: translateX(-5px) translateY(-14px) rotate(-200deg);
+  }
+  img:nth-child(7) {
+    width: 40px;
+    transform: translateX(-10px) translateY(-21px) rotate(250deg);
+  }
+}
+
+@media screen and (max-width: 440px) {
+  .asteroids {
+    flex-direction: row;
+    padding: 1rem 0;
+    box-sizing: border-box;
+
+    &:hover {
+      border-top: 2px solid #ffffff;
+    }
+
+    img:nth-child(1) {
+      width: 45px;
+      transform: translateY(0) translateX(0);
+    }
+    img:nth-child(2) {
+      width: 55px;
+      transform: translateX(0) rotate(220deg);
+    }
+    img:nth-child(3) {
+      width: 70px;
+      transform: translateY(0) translateX(0) rotate(360deg);
+    }
+    img:nth-child(4) {
+      width: 60px;
+      transform: translateY(0) rotate(230deg);
+    }
+    img:nth-child(5) {
+      width: 50px;
+      transform: translateX(0) rotate(330deg);
+    }
+    img:nth-child(6) {
+      width: 40px;
+      transform: translateY(0) translateX(0) rotate(-200deg);
+    }
+    img:nth-child(7) {
+      width: 30px;
+      transform: translateY(0) translateX(0) rotate(250deg);
+    }
+  }
+}
