+++ src/pageContent/explore/Sun.module.css
@@ -1,0 +1,23 @@
+.sun {
+  flex-basis: 10%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 2px solid transparent;
+  transition: 0.5s ease-in-out;
+
+  &:hover {
+    border: 2px solid #ffffff;
+    border-top: 2px solid transparent;
+  }
+
+  img {
+    width: 300px;
+  }
+}
+
+@media screen and (max-width: 440px) {
+  .sun {
+
+  }
+}
