--- src/pageContent/explore/universe/SolarSystem.module.css
@@ -1,106 +1,0 @@
-.system,
-.sun,
-.asteroids {
-  align-self: stretch;
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  box-sizing: border-box;
-  border: 2px solid transparent;
-  transition: 0.5s ease-in-out;
-
-  &:hover {
-    border: 2px solid #ffffff;
-    border-top: 2px solid transparent;
-  }
-}
-
-.sun {
-  flex-basis: 10%;
-
-  img {
-    width: 300px;
-  }
-}
-
-.asteroids {
-  flex-basis: 10%;
-  flex-direction: column;
-  padding: 0 2rem 0 0.5rem;
-
-  img {
-    filter: sepia(1) hue-rotate(5deg);
-  }
-
-  img:nth-child(1) {
-    width: 55px;
-    transform: translateX(-20px) translateY(21px);
-  }
-  img:nth-child(2) {
-    width: 65px;
-    transform: translateY(14px) rotate(220deg);
-  }
-  img:nth-child(3) {
-    width: 80px;
-    transform: translateX(10px) translateY(7px) rotate(360deg);
-  }
-  img:nth-child(4) {
-    width: 70px;
-    transform: translateX(5px) rotate(230deg);
-  }
-  img:nth-child(5) {
-    width: 60px;
-    transform: translateY(-7px) rotate(330deg);
-  }
-  img:nth-child(6) {
-    width: 50px;
-    transform: translateX(-5px) translateY(-14px) rotate(-200deg);
-  }
-  img:nth-child(7) {
-    width: 40px;
-    transform: translateX(-10px) translateY(-21px) rotate(250deg);
-  }
-}
-
-.system {
-  flex-basis: 40%;
-  padding: 0 1rem;
-  gap: 2rem;
-
-  img {
-    height: fit-content;
-  }
-
-  .mercury {
-    width: 24px;
-  }
-  .venus {
-    width: 35px;
-    transform: translateY(-15px);
-  }
-  .earth {
-    width: 45px;
-    transform: translateY(-30px);
-  }
-  .mars {
-    width: 27px;
-    transform: translateY(-40px);
-  }
-  .jupiter {
-    width: 150px;
-    transform: translateY(-50px);
-  }
-  .saturn {
-    width: 200px;
-    margin: 0 -35px;
-    transform: translateY(-30px);
-  }
-  .uranus {
-    width: 100px;
-    transform: translateY(20px);
-  }
-  .neptune {
-    width: 80px;
-    transform: translateY(60px);
-  }
-}
