--- src/pageContent/explore/universe/Universe.module.css
@@ -1,8 +1,0 @@
-.universe {
-  align-self: center;
-  flex: 1;
-  display: flex;
-  align-items: center;
-  overflow: auto;
-  width: 100%;
-}
