+++ src/pages/Explore.tsx
@@ -1,4 +1,4 @@
-import Universe from "@/pageContent/explore/universe/Universe";
+import Universe from "@/pageContent/explore/Universe";
 
 export default function ExplorePage() {
   return <Universe />;
